# backend/swagger.yaml
swagger: "2.0"
info:
  description: "APIs of Logistic Web Service"
  version: "1.0.0"
  title: "Logistic Web Service"
  contact:
    email: "d.bed11@gptgdt.net"
host: "localhost:8000"
basePath: "/api"
schemes:
- "https"
- "http"
...
paths:
...
  /user:
    get:
      tags:
        - "user"
      summary: get user
      produces:
        - application/json
      parameters:
        - in: query
          name: limit
          type: string
          description: limit number of partners 
      responses:
        200:
          description: get user successfully.
          schema:
            type: object
            example: {
              "email": "string",
              "phone": "string",
            }
        400:
          description: partner is not found.
  /user/{userId}:
    get:
      tags:
        - "user"
      summary: get user by id
      produces:
        - application/json
      parameters:
        - in: path
          name: userId
          type: string
          required: true
          description: Id of a user 
      responses:
        200:
          description: get user successfully.
          schema:
            type: object
            example: {
              "name": "string",
            }
        400:
          description: UserId is not found.  
  /user/staff/{userId}:
    put:
      tags:
        - "user"
      summary: "Update a existing staff"
      consumes:
        - multipart/form-data
      parameters:
        - in: path
          name: userId
          schema: 
            type: string
          required: true
          description: user id
        - in: body
          name: user
          description: a existing user.
          schema:
            type: object
            required:
              - name
              - email
              - phone
            properties:
              name:
                type: string
              email:
                type: string
              phone:
                type: string
      requestBody:
        content: 
          multipart/form-data: 
            schema:            
              type: object
              properties:      
                name:           
                  type: string
      responses:
        200:
          description: Update staff successfully.
          schema:
            type: object
            example: {
              "name": "string",
              "email": "string",
              "phone": "string",
            }
        400:
          description: Update staff failed.
...
  /user/customer/{userId}:
    put:
      tags:
        - "user"
      summary: "Update a existing customer"
      consumes:
        - multipart/form-data
      parameters:
        - in: path
          name: userId
          schema: 
            type: string
          required: true
          description: user id
        - in: body
          name: customer
          description: a existing customer.
          schema:
            type: object
            required:
              - name
              - email
              - phone
            properties:
              name:
                type: string
              email:
                type: string
              phone:
                type: string
      requestBody:
        content: 
          multipart/form-data: 
            schema:            
              type: object
              properties:      
                name:           
                  type: string
      responses:
        200:
          description: Update staff successfully.
          schema:
            type: object
            example: {
              "name": "string",
              "email": "string",
              "phone": "string",
            }
        400:
          description: Update staff failed.
  /admin/user/{userId}:  
    delete:
        tags:
          - "user"
        summary: "Delete a existing user"
        produces:
          - application/json
        parameters:
          - in: path
            name: userId
            type: string
            required: true
            description: user Id
        responses:
          200:
            description: Delete user successfully.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      description: Success message
          400:
            description: Update user failed.    
...
  /prohibited-product/{id}:
    get:
      tags:
        - "prohibited-product"
      summary: get prohibited product by id
      produces:
        - application/json
      parameters:
        - in: path
          name: Id
          type: string
          required: true
          description: Id of a prohibited product 
      responses:
        200:
          description: get prohibited product successfully.
          schema:
            type: object
            example: {
            }
        400:
          description: Id is not found. 
  /admin/prohibited-product/create:
    post:
      tags:
        - "prohibited-product"
      summary: "Create a new prohibited product"
      consumes:
        - multipart/form-data
      parameters:
        - in: body
          name: prohibited product
          description: a new prohibited product.
          schema:
            type: object
            required:
              - name
            properties:
              name:           
                  type: string
        - in: formData
      requestBody:
        content: 
          application/json:
            schema:            
              type: object
              properties:      
                name:           
                  type: string
      responses:
        200:
          description: Create prohibited product successfully.
          schema:
            type: object
            example: {
              "name": "string",
            }
        400:
          description: Create prohibited product failed.
  /admin/prohibited-product/{id}:
    put:
      tags:
        - "prohibited-product"
      summary: "Update a existing prohibited product"
      consumes:
        - multipart/form-data
      parameters:
        - in: path
          name: id
          schema: 
            type: string
          required: true
          description: prohibited product id
        - in: body
          name: prohibited product
          description: a existing prohibited product.
          schema:
            type: object
            required:
              - name
            properties:
              name:
                type: string
      requestBody:
        content: 
          multipart/form-data: 
            schema:            
              type: object
              properties:      
                name:           
                  type: string
      responses:
        200:
          description: Update prohibited product successfully.
          schema:
            type: object
            example: {
              "name": "string",
            }
        400:
          description: Update prohibited product failed.
    delete:
      tags:
        - "prohibited-product"
      summary: "Delete a existing prohibited product"
      produces:
        - application/json
      parameter:
        - name: id
          type: string
          required: true
          description: prohibited product Id
      responses:
        200:
          description: Delete prohibited product successfully.
          content:
            application/json:
              schema:
                type: object
                properties: 
                  message:
                    type: string
                    description: Success message
        400: 
          description: Delete prohibited product failed.
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"